Add Maven pom.xml for one of the samples
There's a lot of (apparently quite unecessary) configuration just to
get Maven to look for sources in src/documentation/*. Also there's
duplication of the build/generated-documentation path in
documentation.properties as well as asciidoctor-maven-plugin config.
It's unidiomatic to have to configure that kind of thing in a properties
file anyway for maven, so probably better to allow it to be specified
as a restdocs plugin parameter (when such a thing exists).

But, hey, it works! And the documentation looks pretty.