Improve and simplify quoting for cURL commands
Single quotes are generally safer to use since shells won't try to
interpolate any strings or do variable substitution inside it. This
could be refined to do a check for [&%$] inside the URI/query string
before just blindly surrounding with quotes, but this is safe as it is.

Also, don't escape POST paylods since they are quoted already – this
doesn't work correctly right now.

Closes gh-55