Optional path and req param descriptors should count as documented
Previously, optional path and request parameter descriptors were
ignored when checking that all of the parameters that are present
had been documented. This lead to a false negative if a request
or path parameter was present and was documented with an optional
descriptor.

This commit uses all of the descriptors, not just those that are not
optional, when checking for undocumented parameters.

Closes gh-228