Ensure that additional snippets are only called once
Previously, when additional snippets were added to a
RestDocumentationResultHandler, those snippets would then be called
for every subsequent call to handle(MvcResult). This is problematic as
the snippets and their configuration are specific to a particular
MvcResult that is being documented.

This commit updates RestDocumentationResultHandler so that it clears
its additional snippets each time handle(MvcResult) is called.

Closes gh-243