Re-order table of samples to make it easier for new users
If I'm just writing a simple service with Spring, I'm unlikely
to want to know about restassured. It's great that I have the choice
and I will look at it later, but I'd like to see the more Spring
oriented sampled at the top of the table. And the one that uses Maven.